+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [v0.1.0] - 2026-01-12
 
 - Extend compatibility down to VS Code 1.99.0.
 - Rename `erlab.manager` to `erlab.itool`.